Split JSHash/JSHmac into cell, prototype and constructor files like JSCipher
JSHash.{h,cpp} / JSHmac.{h,cpp} keep the cell class and class-structure setup;
JSHashPrototype/JSHmacPrototype hold the prototype class and its host
functions; JSHashConstructor/JSHmacConstructor hold the InternalFunction and
construct/call. No behavior change.
